+++ ft_strlcpy.c
@@ -6,15 +6,15 @@
 /*   By: aaibar-h <aaibar-h@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/13 09:46:02 by aaibar-h          #+#    #+#             */
-/*   Updated: 2023/01/24 17:53:52 by aaibar-h         ###   ########.fr       */
+/*   Updated: 2023/01/24 18:44:05 by aaibar-h         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlcpy(char *dst, const char *src, int size)
+size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	const int	slen = ft_strlen(src);
+	const size_t	slen = ft_strlen(src);
 
 	if (size == 0)
 		return (slen);
